requirements.txt
# requirements.txt
mlflow>=2.21.2
streamlit==1.44.1
databricks-sql-connector>=3.0.0
databricks-sdk>=0.33.0
pandas>=2.0.0
